+++ routes/web.php
@@ -16,6 +16,7 @@ use App\Elasticsearch\Repositories\ItemRepository;
 use App\Filter\ItemFilter;
 use App\Http\Controllers\Admin\FeaturedArtworkController;
 use App\Http\Controllers\Admin\FeaturedPieceController;
+use App\Http\Controllers\Admin\ItemTagsController;
 use App\Http\Controllers\Admin\ShuffledItemController;
 use App\Http\Controllers\AdminController;
 use App\Http\Controllers\ArticleController;
@@ -45,6 +46,7 @@ use App\Item;
 use App\Notice;
 use App\Order;
 use Illuminate\Support\Facades\Config;
+use Illuminate\Support\Facades\Date;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Facades\Request;
 use Illuminate\Support\Facades\Route;
@@ -226,9 +228,12 @@ function()
         if (empty($item)) {
             abort(404);
         }
+
         $item->timestamps = false;
-        $item->view_count += 1;
-        $item->save();
+        $item->view_count++;
+        $item->last_viewed_at = Date::now();
+        $item->saveQuietly(); // Do not sync to search
+
         $previous = $next = false;
 
         $similar_items = $itemRepository->getSimilar(12, $item)->getCollection();
@@ -357,6 +362,7 @@ Route::group(['middleware' => ['auth', 'can:edit']], function () {
     Route::get('item/search', [ItemController::class, 'search']);
 
     Route::get('item', [ItemController::class, 'index'])->name('item.index');
+    Route::resource('item/tags', ItemTagsController::class)->names('item-tags');
     Route::get('item/{id}/show', [ItemController::class, 'show'])->name('item.show');
     Route::match(['get', 'post'], 'item/create', [ItemController::class, 'create'])->name('item.create');
     Route::match(['get', 'post'], 'item/{id}/edit', [ItemController::class, 'edit'])->name('item.edit');