+++ CHANGELOG.md
@@ -3,6 +3,18 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 
+## [2.61.0] - 2022-07-29
+### Added
+- Mass tagging feature in admin
+
+### Changed
+- KolekciaController to optimize queries
+- Item view count is indexed every 24 hours instead of immediately
+
+### Fixed
+- sorting of collection items
+- Item admin form to not show 'item.' prefix in choices
+
 ## [2.60.0] - 2022-07-22
 ### Changed
 - Laravel version to 9